Renamed FormDownloadList to FormDownloadListActivity
