+++ collect_app/src/main/java/org/odk/collect/android/tasks/ServerPollingJob.java
@@ -27,7 +27,7 @@ import com.evernote.android.job.JobManager;
 import com.evernote.android.job.JobRequest;
 
 import org.odk.collect.android.R;
-import org.odk.collect.android.activities.FormDownloadList;
+import org.odk.collect.android.activities.FormDownloadListActivity;
 import org.odk.collect.android.activities.NotificationActivity;
 import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.dao.FormsDao;
@@ -46,7 +46,7 @@ import java.util.Map;
 
 import javax.inject.Inject;
 
-import static org.odk.collect.android.activities.FormDownloadList.DISPLAY_ONLY_UPDATED_FORMS;
+import static org.odk.collect.android.activities.FormDownloadListActivity.DISPLAY_ONLY_UPDATED_FORMS;
 import static org.odk.collect.android.preferences.GeneralKeys.KEY_AUTOMATIC_UPDATE;
 import static org.odk.collect.android.provider.FormsProviderAPI.FormsColumns.JR_FORM_ID;
 import static org.odk.collect.android.provider.FormsProviderAPI.FormsColumns.LAST_DETECTED_FORM_VERSION_HASH;
@@ -157,7 +157,7 @@ public class ServerPollingJob extends Job {
     }
 
     private void informAboutNewAvailableForms() {
-        Intent intent = new Intent(getContext(), FormDownloadList.class);
+        Intent intent = new Intent(getContext(), FormDownloadListActivity.class);
         intent.putExtra(DISPLAY_ONLY_UPDATED_FORMS, true);
         PendingIntent contentIntent = PendingIntent.getActivity(getContext(), FORM_UPDATES_AVAILABLE_NOTIFICATION, intent, PendingIntent.FLAG_UPDATE_CURRENT);
 
@@ -171,7 +171,7 @@ public class ServerPollingJob extends Job {
     private void informAboutNewDownloadedForms(String title, HashMap<FormDetails, String> result) {
         Intent intent = new Intent(Collect.getInstance(), NotificationActivity.class);
         intent.putExtra(NotificationActivity.NOTIFICATION_TITLE, title);
-        intent.putExtra(NotificationActivity.NOTIFICATION_MESSAGE, FormDownloadList.getDownloadResultMessage(result));
+        intent.putExtra(NotificationActivity.NOTIFICATION_MESSAGE, FormDownloadListActivity.getDownloadResultMessage(result));
         PendingIntent contentIntent = PendingIntent.getActivity(getContext(), FORMS_DOWNLOADED_NOTIFICATION, intent, PendingIntent.FLAG_UPDATE_CURRENT);
 
         NotificationUtils.showNotification(contentIntent,
